Refactored Code: Made code more modular, Checked spelling, and formatted with Prettier
