+++ app/routes/index.js
@@ -1,0 +1,8 @@
+// Import express and router
+const express = require("express");
+const router = express.Router();
+
+// DEFINE AND USE Routes
+
+// export the router
+// module.exports = router;
