--- routes/index.js
@@ -1,25 +1,0 @@
-// Import express and router
-const express = require("express");
-const router = express.Router();
-
-// Import Spotify Module
-
-// Actions:
-// Login
-// Auth
-// Search
-
-// Router Use (Endpoint Routes)
-//router.use("/endpoint, routesVariable");
-
-// Handle the GET root route
-// router.get("/", (req, res) => {
-//   // Respond with a JSON message
-//   res.status(200).json({
-//     success: true,
-//     message: ` ${req.method} - Request made`,
-//   });
-// });
-
-// export the router
-// module.exports = router;
