+++ server.js
@@ -1,0 +1,20 @@
+// Importing the express module
+const app = require("./app");
+
+// Importing the environment variable configuration
+require("dotenv").config();
+
+// Import the database
+const connectDB = require("./app/db/config");
+
+// Run the database (MongoDB) asynchronous function
+connectDB();
+
+//  Port Variable from .env file
+const PORT = process.env.PORT || 3000;
+
+// Starting the server and listening on the specified port
+// listening for incoming requests
+app.listen(PORT, () => {
+  console.log(`Server is running on ${PORT} `);
+});
